Segmented probe logic by type to increase readability
